Move rich-text-editor-utils into a non-exported folder
