+++ packages/nimble-components/src/rich-text/viewer/tests/rich-text-viewer.spec.ts
@@ -11,7 +11,7 @@ import {
 import { richTextMentionUsersViewTag } from '../../../rich-text-mention/users/view';
 import { MappingUser, mappingUserTag } from '../../../mapping/user';
 import type { MappingConfiguration } from '../../editor/testing/types';
-import type { UserMentionElements } from '../../editor/testing/rich-text-editor-utils';
+import type { UserMentionElements } from '../../editor/models/tests/rich-text-editor-utils';
 
 async function setup(): Promise<Fixture<RichTextViewer>> {
     return await fixture<RichTextViewer>(
